Move cache-key-func to slurm launcher
